+++ src/components/app/app.jsx
@@ -17,7 +17,6 @@ import { getUser } from "../../utils/api";
 import { CHECK_TOKEN, GET_USER, getIngredients } from "../../services/actions";
 import { OrderPage } from "../../pages/orders/orders";
 import { FeedPage } from "../../pages/feed/feed";
-import { OrderDescription } from "../order-description/order-description";
 
 
 
@@ -55,7 +54,7 @@ function App() {
       <AppHeader />
       <Routes>
         <Route path="/feed" element={<FeedPage />} />
-        <Route path="/feed/:id" element={<OrderDescription />} />
+        {/* <Route path="/feed/:id" element={<OrderDescription />} /> */}
         <Route path="" element={<Home modal={modal} setModal={setModal} />} />
         <Route path="/login" element={<OnlyUnAuth component={<LoginPage />} />} />
         <Route path="/register" element={<OnlyUnAuth component={<RegistrationPage />} />} />
