+++ src/components/order-description/order-description.module.css
@@ -1,0 +1,137 @@
+.main {
+  width: 1280px;
+  margin: 40px auto 0;
+  display: flex;
+  flex-direction: column;
+}
+
+.box {
+  display: flex;
+  margin-top: 20px;
+}
+
+.order_list {
+  width: 584px;
+  height: 90wh;
+}
+
+.order_numbers {
+  display: flex;
+  column-gap: 36px;
+}
+
+.order_number {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.order {
+  display: flex;
+  flex-direction: column;
+  width: 536px;
+  height: 166px;
+  padding: 24px;
+  border-radius: 40px;
+  background-color: #1C1C21;
+  margin-bottom: 16px;
+  row-gap: 24px;
+
+}
+
+.ingredient_image {
+  width: 56px;
+  height: 56px;
+  margin-left: -12px;
+  border-radius: 50%;
+  border: 1px solid #4C4CFF;
+  object-fit: cover;
+  background-color: #1C1C21;
+}
+
+.ingredients {
+  display: flex;
+  margin-left: 12px;
+}
+
+
+.ingredient_image:nth-child(1) {
+  z-index: 7;
+}
+
+.ingredient_image:nth-child(2) {
+  z-index: 6;
+}
+
+.ingredient_image:nth-child(3) {
+  z-index: 5;
+}
+
+.ingredient_image:nth-child(4) {
+  z-index: 4;
+}
+
+.ingredient_image:nth-child(5) {
+  z-index: 3;
+}
+
+.ingredient_image:nth-child(6) {
+  z-index: 1;
+}
+
+.done_order {
+  color: #0CC;
+}
+
+.done_orders {
+  width: 272px;
+  margin-top: 24px;
+  display: flex;
+  flex-direction: column;
+  row-gap: 8px;
+}
+
+.total_price {
+  display: flex;
+  column-gap: 8px;
+}
+
+.box_ingredients {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.orders {
+  margin-left: 60px;
+  display: flex;
+  flex-direction: column;
+  row-gap: 60px;
+}
+
+.shadow {
+  text-shadow: rgb(51, 51, 255) 1px 0 10px;
+}
+
+.layer {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 56px;
+  height: 56px;
+  border-radius: 50%;
+  opacity: 0.6;
+  background-color: #1C1C21;
+  z-index: 2;
+  margin: 0 0 0 -12px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.hidden_elements {
+  position: relative;
+  width: 56px;
+  height: 56px;
+  border-radius: 50%;
+}
