+++ src/components/order-ingredient/order-ingredient.jsx
@@ -1,0 +1,7 @@
+import style from './order-ingredient.module.css';
+
+export const OrderIngredient = ({ card }) => {
+  return (
+    <img className={style.ingredient_image} src={card.image_mobile} />
+  )
+}
