+++ src/components/order-ingredient/order-ingredient.module.css
@@ -1,0 +1,39 @@
+.ingredient_image {
+  width: 56px;
+  height: 56px;
+  margin-left: -12px;
+  border-radius: 50%;
+  border: 1px solid #4C4CFF;
+  object-fit: cover;
+  background-color: #1C1C21;
+}
+
+.ingredients {
+  display: flex;
+  margin-left: 12px;
+}
+
+
+.ingredient_image:nth-child(1) {
+  z-index: 7;
+}
+
+.ingredient_image:nth-child(2) {
+  z-index: 6;
+}
+
+.ingredient_image:nth-child(3) {
+  z-index: 5;
+}
+
+.ingredient_image:nth-child(4) {
+  z-index: 4;
+}
+
+.ingredient_image:nth-child(5) {
+  z-index: 3;
+}
+
+.ingredient_image:nth-child(6) {
+  z-index: 1;
+}
