+++ src/pages/feed/feed.module.css
@@ -39,46 +39,6 @@
 
 }
 
-.ingredient_image {
-  width: 56px;
-  height: 56px;
-  margin-left: -12px;
-  border-radius: 50%;
-  border: 1px solid #4C4CFF;
-  object-fit: cover;
-  background-color: #1C1C21;
-}
-
-.ingredients {
-  display: flex;
-  margin-left: 12px;
-}
-
-
-.ingredient_image:nth-child(1) {
-  z-index: 7;
-}
-
-.ingredient_image:nth-child(2) {
-  z-index: 6;
-}
-
-.ingredient_image:nth-child(3) {
-  z-index: 5;
-}
-
-.ingredient_image:nth-child(4) {
-  z-index: 4;
-}
-
-.ingredient_image:nth-child(5) {
-  z-index: 3;
-}
-
-.ingredient_image:nth-child(6) {
-  z-index: 1;
-}
-
 .done_order {
   color: #0CC;
 }
@@ -91,16 +51,6 @@
   row-gap: 8px;
 }
 
-.total_price {
-  display: flex;
-  column-gap: 8px;
-}
-
-.box_ingredients {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
 
 .orders {
   margin-left: 60px;
@@ -111,27 +61,4 @@
 
 .shadow {
   text-shadow: rgb(51, 51, 255) 1px 0 10px;
-}
-
-.layer {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 56px;
-  height: 56px;
-  border-radius: 50%;
-  opacity: 0.6;
-  background-color: #1C1C21;
-  z-index: 2;
-  margin: 0 0 0 -12px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.hidden_elements {
-  position: relative;
-  width: 56px;
-  height: 56px;
-  border-radius: 50%;
 }
